+++ packages/common/package.json
@@ -16,6 +16,7 @@
     "dompurify": "^2.4.3",
     "ethers": "^5.6.5",
     "framer-motion": "^10.12.7",
+    "grants-stack-data-client": "github:gitcoinco/grants-stack-data-client",
     "markdown-it": "^13.0.1",
     "react": "^18.2.0",
     "react-hook-form": "^7.42.1",
