+++ packages/grant-explorer/src/index.tsx
@@ -11,6 +11,9 @@ import { initSentry } from "./sentry";
 import { initTagmanager } from "./tagmanager";
 import { chains, config } from "./app/wagmi";
 import reportWebVitals from "./reportWebVitals";
+import { GrantsStackDataProvider } from "common/src/grantsStackDataClientContext";
+import { GrantsStackDataClient } from "grants-stack-data-client";
+import { getConfig } from "common/src/config";
 
 import "./index.css";
 
@@ -41,49 +44,62 @@ const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
 );
 
+const GRANTS_STACK_DATA_APPLICATIONS_PAGE_SIZE = 50;
+
+const grantsStackDataClient = new GrantsStackDataClient({
+  baseUrl: getConfig().grantsStackDataClient.baseUrl,
+  applications: {
+    pagination: {
+      pageSize: GRANTS_STACK_DATA_APPLICATIONS_PAGE_SIZE,
+    },
+  },
+});
+
 root.render(
   <React.StrictMode>
     <ChakraProvider>
       <WagmiConfig config={config}>
         <RainbowKitProvider coolMode chains={chains}>
           <RoundProvider>
-            <HashRouter>
-              <Routes>
-                {/* Protected Routes */}
-                <Route element={<Auth />} />
+            <GrantsStackDataProvider client={grantsStackDataClient}>
+              <HashRouter>
+                <Routes>
+                  {/* Protected Routes */}
+                  <Route element={<Auth />} />
 
-                {/* Default Route */}
-                <Route path="/" element={<LandingPage />} />
+                  {/* Default Route */}
+                  <Route path="/" element={<LandingPage />} />
 
-                <Route path="/rounds" element={<ExploreRoundsPage />} />
-                <Route path="/projects" element={<ProjectsPage />} />
+                  <Route path="/rounds" element={<ExploreRoundsPage />} />
+                  <Route path="/projects" element={<ProjectsPage />} />
 
-                {/* Round Routes */}
-                <Route
-                  path="/round/:chainId/:roundId"
-                  element={<ViewRound />}
-                />
-                <Route
-                  path="/round/:chainId/:roundId/:applicationId"
-                  element={<ViewProjectDetails />}
-                />
+                  {/* Round Routes */}
+                  <Route
+                    path="/round/:chainId/:roundId"
+                    element={<ViewRound />}
+                  />
+                  <Route
+                    path="/round/:chainId/:roundId/:applicationId"
+                    element={<ViewProjectDetails />}
+                  />
 
-                <Route path="/cart" element={<ViewCart />} />
+                  <Route path="/cart" element={<ViewCart />} />
 
-                <Route path="/thankyou" element={<ThankYou />} />
+                  <Route path="/thankyou" element={<ThankYou />} />
 
-                <Route
-                  path="/contributors/:address"
-                  element={<ViewContributionHistory />}
-                />
+                  <Route
+                    path="/contributors/:address"
+                    element={<ViewContributionHistory />}
+                  />
 
-                {/* Access Denied */}
-                <Route path="/access-denied" element={<AccessDenied />} />
+                  {/* Access Denied */}
+                  <Route path="/access-denied" element={<AccessDenied />} />
 
-                {/* 404 */}
-                <Route path="*" element={<NotFound />} />
-              </Routes>
-            </HashRouter>
+                  {/* 404 */}
+                  <Route path="*" element={<NotFound />} />
+                </Routes>
+              </HashRouter>
+            </GrantsStackDataProvider>
           </RoundProvider>
         </RainbowKitProvider>
       </WagmiConfig>
